backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]>=0.20.0
fastapi-amis-admin>=0.7.1
fastapi-user-auth>=0.7.3
casbin==1.43.0
sqlalchemy[asyncio]>=2.0.0
sqlmodel==0.0.18
aioodbc>=0.5.0
pyodbc>=4.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
Babel>=2.12.0
aiofiles>=23.0.0
python-multipart>=0.0.6
greenlet>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
